requirements_segmentation.txt
openai>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0

